fix(inference): refactor local adapter lifecycle
