+++ src/lib/inference/ollama/proxy.ts
@@ -5,8 +5,6 @@
 // Ollama auth-proxy lifecycle: token persistence, PID management,
 // proxy start/stop, model pull and validation.
 
-const fs = require("fs");
-const os = require("os");
 const path = require("path");
 const { spawn, spawnSync } = require("child_process");
 const http = require("http");
@@ -30,10 +28,21 @@ const {
   formatOllamaProxyUnreachableMessage,
   probeOllamaProxySandboxReachability,
 } = require("../../onboard/ollama-proxy-reachability");
+const {
+  DEFAULT_LOCAL_ADAPTER_STATE_DIR,
+  isLocalAdapterProcess,
+  killLocalAdapterPid,
+  loadLocalAdapterPid,
+  persistLocalAdapterPid,
+  readLocalAdapterTextFile,
+  removeLocalAdapterFile,
+  spawnDetachedNodeAdapter,
+  writeLocalAdapterSecretFile,
+} = require("../local-adapter-lifecycle");
 
 // ── State ────────────────────────────────────────────────────────
 
-const PROXY_STATE_DIR = path.join(os.homedir(), ".nemoclaw");
+const PROXY_STATE_DIR = DEFAULT_LOCAL_ADAPTER_STATE_DIR;
 const PROXY_TOKEN_PATH = path.join(PROXY_STATE_DIR, "ollama-proxy-token");
 const PROXY_PID_PATH = path.join(PROXY_STATE_DIR, "ollama-auth-proxy.pid");
 
@@ -43,21 +52,10 @@ function sleep(seconds) {
   spawnSync("sleep", [String(seconds)]);
 }
 
-// ── Proxy state dir ──────────────────────────────────────────────
-
-function ensureProxyStateDir(): void {
-  if (!fs.existsSync(PROXY_STATE_DIR)) {
-    fs.mkdirSync(PROXY_STATE_DIR, { recursive: true });
-  }
-}
-
 // ── Token persistence ────────────────────────────────────────────
 
 function persistProxyToken(token: string): void {
-  ensureProxyStateDir();
-  fs.writeFileSync(PROXY_TOKEN_PATH, token, { mode: 0o600 });
-  // mode only applies on creation; ensure permissions on existing files too
-  fs.chmodSync(PROXY_TOKEN_PATH, 0o600);
+  writeLocalAdapterSecretFile(PROXY_TOKEN_PATH, token);
 }
 
 // Persist the proxy token then probe sandbox → proxy reachability. Runs
@@ -75,15 +73,7 @@ async function persistAndProbeOllamaProxy(token: string): Promise<void> {
 }
 
 function loadPersistedProxyToken(): string | null {
-  try {
-    if (fs.existsSync(PROXY_TOKEN_PATH)) {
-      const token = fs.readFileSync(PROXY_TOKEN_PATH, "utf-8").trim();
-      return token || null;
-    }
-  } catch {
-    /* ignore */
-  }
-  return null;
+  return readLocalAdapterTextFile(PROXY_TOKEN_PATH);
 }
 
 function curlAuthHeaderConfig(token: string): string {
@@ -122,63 +112,45 @@ function runCurlCaptureWithAuthConfig(args: string[], endpoint: string, token: s
 // ── PID persistence ──────────────────────────────────────────────
 
 function persistProxyPid(pid: number | null | undefined): void {
-  if (!Number.isInteger(pid) || pid <= 0) return;
-  ensureProxyStateDir();
-  fs.writeFileSync(PROXY_PID_PATH, `${pid}\n`, { mode: 0o600 });
-  fs.chmodSync(PROXY_PID_PATH, 0o600);
+  persistLocalAdapterPid(PROXY_PID_PATH, pid);
 }
 
 function loadPersistedProxyPid(): number | null {
-  try {
-    if (!fs.existsSync(PROXY_PID_PATH)) return null;
-    const raw = fs.readFileSync(PROXY_PID_PATH, "utf-8").trim();
-    const pid = Number.parseInt(raw, 10);
-    return Number.isInteger(pid) && pid > 0 ? pid : null;
-  } catch {
-    return null;
-  }
+  return loadLocalAdapterPid(PROXY_PID_PATH);
 }
 
 function clearPersistedProxyPid(): void {
-  try {
-    if (fs.existsSync(PROXY_PID_PATH)) {
-      fs.unlinkSync(PROXY_PID_PATH);
-    }
-  } catch {
-    /* ignore */
-  }
+  removeLocalAdapterFile(PROXY_PID_PATH);
 }
 
 // ── Process management ───────────────────────────────────────────
 
 function isOllamaProxyProcess(pid: number | null | undefined): boolean {
-  if (!Number.isInteger(pid) || pid <= 0) return false;
-  const cmdline = runCapture(["ps", "-p", String(pid), "-o", "args="], { ignoreError: true });
-  return Boolean(cmdline && cmdline.includes("ollama-auth-proxy.js"));
+  return isLocalAdapterProcess(pid, "ollama-auth-proxy.js", runCapture);
 }
 
 function spawnOllamaAuthProxy(token: string): number | null {
-  const child = spawn(process.execPath, [path.join(SCRIPTS, "ollama-auth-proxy.js")], {
-    detached: true,
-    stdio: "ignore",
-    env: buildSubprocessEnv({
+  const child = spawnDetachedNodeAdapter({
+    scriptPath: path.join(SCRIPTS, "ollama-auth-proxy.js"),
+    env: {
       OLLAMA_PROXY_TOKEN: token,
       OLLAMA_PROXY_PORT: String(OLLAMA_PROXY_PORT),
       OLLAMA_BACKEND_PORT: String(OLLAMA_PORT),
-    }),
+    },
+    buildEnv: buildSubprocessEnv,
   });
-  child.unref();
   persistProxyPid(child.pid);
   return child.pid ?? null;
 }
 
 function killStaleProxy(): void {
   try {
-    const persistedPid = loadPersistedProxyPid();
-    if (isOllamaProxyProcess(persistedPid)) {
-      run(["kill", String(persistedPid)], { ignoreError: true, suppressOutput: true });
-    }
-    clearPersistedProxyPid();
+    killLocalAdapterPid({
+      pidPath: PROXY_PID_PATH,
+      processNeedle: "ollama-auth-proxy.js",
+      run,
+      runCapture,
+    });
 
     // Best-effort cleanup for older proxy processes created before the PID file
     // existed. Only kill processes that are actually the auth proxy, not
